@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: dogstatsd/dogstatsd.go
+//
+// Generated by this command:
+//
+//	mockgen -source=dogstatsd/dogstatsd.go -destination=dogstatsd/mocks/dogstatsd.go -package mocks
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -8,13 +13,14 @@ import (
 	reflect "reflect"
 	time "time"
 
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockClient is a mock of Client interface.
 type MockClient struct {
 	ctrl     *gomock.Controller
 	recorder *MockClientMockRecorder
+	isgomock struct{}
 }
 
 // MockClientMockRecorder is the mock recorder for MockClient.
@@ -43,7 +49,7 @@ func (m *MockClient) Count(name string, value int64, tags []string, rate float64
 }
 
 // Count indicates an expected call of Count.
-func (mr *MockClientMockRecorder) Count(name, value, tags, rate interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) Count(name, value, tags, rate any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Count", reflect.TypeOf((*MockClient)(nil).Count), name, value, tags, rate)
 }
@@ -57,7 +63,7 @@ func (m *MockClient) Distribution(name string, value float64, tags []string, rat
 }
 
 // Distribution indicates an expected call of Distribution.
-func (mr *MockClientMockRecorder) Distribution(name, value, tags, rate interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) Distribution(name, value, tags, rate any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Distribution", reflect.TypeOf((*MockClient)(nil).Distribution), name, value, tags, rate)
 }
@@ -71,7 +77,7 @@ func (m *MockClient) Gauge(name string, value float64, tags []string, rate float
 }
 
 // Gauge indicates an expected call of Gauge.
-func (mr *MockClientMockRecorder) Gauge(name, value, tags, rate interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) Gauge(name, value, tags, rate any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Gauge", reflect.TypeOf((*MockClient)(nil).Gauge), name, value, tags, rate)
 }
@@ -85,7 +91,7 @@ func (m *MockClient) Histogram(name string, value float64, tags []string, rate f
 }
 
 // Histogram indicates an expected call of Histogram.
-func (mr *MockClientMockRecorder) Histogram(name, value, tags, rate interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) Histogram(name, value, tags, rate any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Histogram", reflect.TypeOf((*MockClient)(nil).Histogram), name, value, tags, rate)
 }
@@ -99,7 +105,7 @@ func (m *MockClient) Incr(name string, tags []string, rate float64) error {
 }
 
 // Incr indicates an expected call of Incr.
-func (mr *MockClientMockRecorder) Incr(name, tags, rate interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) Incr(name, tags, rate any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Incr", reflect.TypeOf((*MockClient)(nil).Incr), name, tags, rate)
 }
@@ -113,7 +119,7 @@ func (m *MockClient) Timing(name string, value time.Duration, tags []string, rat
 }
 
 // Timing indicates an expected call of Timing.
-func (mr *MockClientMockRecorder) Timing(name, value, tags, rate interface{}) *gomock.Call {
+func (mr *MockClientMockRecorder) Timing(name, value, tags, rate any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Timing", reflect.TypeOf((*MockClient)(nil).Timing), name, value, tags, rate)
 }
