@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: pg/interfaces/pg.go
+//
+// Generated by this command:
+//
+//	mockgen -source=pg/interfaces/pg.go -destination=pg/mocks/pg.go -package mocks
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -9,16 +14,17 @@ import (
 	io "io"
 	reflect "reflect"
 
-	pg "github.com/go-pg/pg"
-	orm "github.com/go-pg/pg/orm"
-	gomock "github.com/golang/mock/gomock"
+	pg "github.com/go-pg/pg/v10"
+	orm "github.com/go-pg/pg/v10/orm"
 	interfaces "github.com/topfreegames/extensions/v9/pg/interfaces"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockQueryable is a mock of Queryable interface.
 type MockQueryable struct {
 	ctrl     *gomock.Controller
 	recorder *MockQueryableMockRecorder
+	isgomock struct{}
 }
 
 // MockQueryableMockRecorder is the mock recorder for MockQueryable.
@@ -39,9 +45,9 @@ func (m *MockQueryable) EXPECT() *MockQueryableMockRecorder {
 }
 
 // Exec mocks base method.
-func (m *MockQueryable) Exec(arg0 interface{}, arg1 ...interface{}) (orm.Result, error) {
+func (m *MockQueryable) Exec(arg0 any, arg1 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0}
+	varargs := []any{arg0}
 	for _, a := range arg1 {
 		varargs = append(varargs, a)
 	}
@@ -52,16 +58,36 @@ func (m *MockQueryable) Exec(arg0 interface{}, arg1 ...interface{}) (orm.Result,
 }
 
 // Exec indicates an expected call of Exec.
-func (mr *MockQueryableMockRecorder) Exec(arg0 interface{}, arg1 ...interface{}) *gomock.Call {
+func (mr *MockQueryableMockRecorder) Exec(arg0 any, arg1 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0}, arg1...)
+	varargs := append([]any{arg0}, arg1...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Exec", reflect.TypeOf((*MockQueryable)(nil).Exec), varargs...)
 }
 
-// ExecOne mocks base method.
-func (m *MockQueryable) ExecOne(arg0 interface{}, arg1 ...interface{}) (orm.Result, error) {
+// ExecContext mocks base method.
+func (m *MockQueryable) ExecContext(arg0 context.Context, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0}
+	varargs := []any{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ExecContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ExecContext indicates an expected call of ExecContext.
+func (mr *MockQueryableMockRecorder) ExecContext(arg0, arg1 any, arg2 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecContext", reflect.TypeOf((*MockQueryable)(nil).ExecContext), varargs...)
+}
+
+// ExecOne mocks base method.
+func (m *MockQueryable) ExecOne(arg0 any, arg1 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0}
 	for _, a := range arg1 {
 		varargs = append(varargs, a)
 	}
@@ -72,16 +98,36 @@ func (m *MockQueryable) ExecOne(arg0 interface{}, arg1 ...interface{}) (orm.Resu
 }
 
 // ExecOne indicates an expected call of ExecOne.
-func (mr *MockQueryableMockRecorder) ExecOne(arg0 interface{}, arg1 ...interface{}) *gomock.Call {
+func (mr *MockQueryableMockRecorder) ExecOne(arg0 any, arg1 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0}, arg1...)
+	varargs := append([]any{arg0}, arg1...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecOne", reflect.TypeOf((*MockQueryable)(nil).ExecOne), varargs...)
 }
 
-// Model mocks base method.
-func (m *MockQueryable) Model(model ...interface{}) *orm.Query {
+// ExecOneContext mocks base method.
+func (m *MockQueryable) ExecOneContext(arg0 context.Context, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ExecOneContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ExecOneContext indicates an expected call of ExecOneContext.
+func (mr *MockQueryableMockRecorder) ExecOneContext(arg0, arg1 any, arg2 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecOneContext", reflect.TypeOf((*MockQueryable)(nil).ExecOneContext), varargs...)
+}
+
+// Model mocks base method.
+func (m *MockQueryable) Model(model ...any) *orm.Query {
+	m.ctrl.T.Helper()
+	varargs := []any{}
 	for _, a := range model {
 		varargs = append(varargs, a)
 	}
@@ -91,15 +137,15 @@ func (m *MockQueryable) Model(model ...interface{}) *orm.Query {
 }
 
 // Model indicates an expected call of Model.
-func (mr *MockQueryableMockRecorder) Model(model ...interface{}) *gomock.Call {
+func (mr *MockQueryableMockRecorder) Model(model ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Model", reflect.TypeOf((*MockQueryable)(nil).Model), model...)
 }
 
 // Query mocks base method.
-func (m *MockQueryable) Query(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result, error) {
+func (m *MockQueryable) Query(arg0, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1}
+	varargs := []any{arg0, arg1}
 	for _, a := range arg2 {
 		varargs = append(varargs, a)
 	}
@@ -110,16 +156,36 @@ func (m *MockQueryable) Query(arg0, arg1 interface{}, arg2 ...interface{}) (orm.
 }
 
 // Query indicates an expected call of Query.
-func (mr *MockQueryableMockRecorder) Query(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+func (mr *MockQueryableMockRecorder) Query(arg0, arg1 any, arg2 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	varargs := append([]any{arg0, arg1}, arg2...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Query", reflect.TypeOf((*MockQueryable)(nil).Query), varargs...)
 }
 
-// QueryOne mocks base method.
-func (m *MockQueryable) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result, error) {
+// QueryContext mocks base method.
+func (m *MockQueryable) QueryContext(arg0 context.Context, arg1, arg2 any, arg3 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1}
+	varargs := []any{arg0, arg1, arg2}
+	for _, a := range arg3 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "QueryContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// QueryContext indicates an expected call of QueryContext.
+func (mr *MockQueryableMockRecorder) QueryContext(arg0, arg1, arg2 any, arg3 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1, arg2}, arg3...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryContext", reflect.TypeOf((*MockQueryable)(nil).QueryContext), varargs...)
+}
+
+// QueryOne mocks base method.
+func (m *MockQueryable) QueryOne(arg0, arg1 any, arg2 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1}
 	for _, a := range arg2 {
 		varargs = append(varargs, a)
 	}
@@ -130,16 +196,37 @@ func (m *MockQueryable) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) (o
 }
 
 // QueryOne indicates an expected call of QueryOne.
-func (mr *MockQueryableMockRecorder) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+func (mr *MockQueryableMockRecorder) QueryOne(arg0, arg1 any, arg2 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	varargs := append([]any{arg0, arg1}, arg2...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryOne", reflect.TypeOf((*MockQueryable)(nil).QueryOne), varargs...)
+}
+
+// QueryOneContext mocks base method.
+func (m *MockQueryable) QueryOneContext(arg0 context.Context, arg1, arg2 any, arg3 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1, arg2}
+	for _, a := range arg3 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "QueryOneContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// QueryOneContext indicates an expected call of QueryOneContext.
+func (mr *MockQueryableMockRecorder) QueryOneContext(arg0, arg1, arg2 any, arg3 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1, arg2}, arg3...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryOneContext", reflect.TypeOf((*MockQueryable)(nil).QueryOneContext), varargs...)
 }
 
 // MockORM is a mock of ORM interface.
 type MockORM struct {
 	ctrl     *gomock.Controller
 	recorder *MockORMMockRecorder
+	isgomock struct{}
 }
 
 // MockORMMockRecorder is the mock recorder for MockORM.
@@ -160,9 +247,9 @@ func (m *MockORM) EXPECT() *MockORMMockRecorder {
 }
 
 // CopyFrom mocks base method.
-func (m *MockORM) CopyFrom(r io.Reader, query interface{}, params ...interface{}) (orm.Result, error) {
+func (m *MockORM) CopyFrom(r io.Reader, query any, params ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{r, query}
+	varargs := []any{r, query}
 	for _, a := range params {
 		varargs = append(varargs, a)
 	}
@@ -173,16 +260,16 @@ func (m *MockORM) CopyFrom(r io.Reader, query interface{}, params ...interface{}
 }
 
 // CopyFrom indicates an expected call of CopyFrom.
-func (mr *MockORMMockRecorder) CopyFrom(r, query interface{}, params ...interface{}) *gomock.Call {
+func (mr *MockORMMockRecorder) CopyFrom(r, query any, params ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{r, query}, params...)
+	varargs := append([]any{r, query}, params...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CopyFrom", reflect.TypeOf((*MockORM)(nil).CopyFrom), varargs...)
 }
 
 // CopyTo mocks base method.
-func (m *MockORM) CopyTo(w io.Writer, query interface{}, params ...interface{}) (orm.Result, error) {
+func (m *MockORM) CopyTo(w io.Writer, query any, params ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{w, query}
+	varargs := []any{w, query}
 	for _, a := range params {
 		varargs = append(varargs, a)
 	}
@@ -193,44 +280,46 @@ func (m *MockORM) CopyTo(w io.Writer, query interface{}, params ...interface{}) 
 }
 
 // CopyTo indicates an expected call of CopyTo.
-func (mr *MockORMMockRecorder) CopyTo(w, query interface{}, params ...interface{}) *gomock.Call {
+func (mr *MockORMMockRecorder) CopyTo(w, query any, params ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{w, query}, params...)
+	varargs := append([]any{w, query}, params...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CopyTo", reflect.TypeOf((*MockORM)(nil).CopyTo), varargs...)
 }
 
 // Delete mocks base method.
-func (m *MockORM) Delete(model interface{}) error {
+func (m *MockORM) Delete(model any) (orm.Result, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Delete", model)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // Delete indicates an expected call of Delete.
-func (mr *MockORMMockRecorder) Delete(model interface{}) *gomock.Call {
+func (mr *MockORMMockRecorder) Delete(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Delete", reflect.TypeOf((*MockORM)(nil).Delete), model)
 }
 
 // ForceDelete mocks base method.
-func (m *MockORM) ForceDelete(model interface{}) error {
+func (m *MockORM) ForceDelete(model any) (orm.Result, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "ForceDelete", model)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // ForceDelete indicates an expected call of ForceDelete.
-func (mr *MockORMMockRecorder) ForceDelete(model interface{}) *gomock.Call {
+func (mr *MockORMMockRecorder) ForceDelete(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ForceDelete", reflect.TypeOf((*MockORM)(nil).ForceDelete), model)
 }
 
 // FormatQuery mocks base method.
-func (m *MockORM) FormatQuery(b []byte, query string, params ...interface{}) []byte {
+func (m *MockORM) FormatQuery(b []byte, query string, params ...any) []byte {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{b, query}
+	varargs := []any{b, query}
 	for _, a := range params {
 		varargs = append(varargs, a)
 	}
@@ -240,32 +329,33 @@ func (m *MockORM) FormatQuery(b []byte, query string, params ...interface{}) []b
 }
 
 // FormatQuery indicates an expected call of FormatQuery.
-func (mr *MockORMMockRecorder) FormatQuery(b, query interface{}, params ...interface{}) *gomock.Call {
+func (mr *MockORMMockRecorder) FormatQuery(b, query any, params ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{b, query}, params...)
+	varargs := append([]any{b, query}, params...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "FormatQuery", reflect.TypeOf((*MockORM)(nil).FormatQuery), varargs...)
 }
 
 // Insert mocks base method.
-func (m *MockORM) Insert(model ...interface{}) error {
+func (m *MockORM) Insert(model ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{}
 	for _, a := range model {
 		varargs = append(varargs, a)
 	}
 	ret := m.ctrl.Call(m, "Insert", varargs...)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // Insert indicates an expected call of Insert.
-func (mr *MockORMMockRecorder) Insert(model ...interface{}) *gomock.Call {
+func (mr *MockORMMockRecorder) Insert(model ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Insert", reflect.TypeOf((*MockORM)(nil).Insert), model...)
 }
 
 // Select mocks base method.
-func (m *MockORM) Select(model interface{}) error {
+func (m *MockORM) Select(model any) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Select", model)
 	ret0, _ := ret[0].(error)
@@ -273,21 +363,22 @@ func (m *MockORM) Select(model interface{}) error {
 }
 
 // Select indicates an expected call of Select.
-func (mr *MockORMMockRecorder) Select(model interface{}) *gomock.Call {
+func (mr *MockORMMockRecorder) Select(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Select", reflect.TypeOf((*MockORM)(nil).Select), model)
 }
 
 // Update mocks base method.
-func (m *MockORM) Update(model interface{}) error {
+func (m *MockORM) Update(model any) (orm.Result, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Update", model)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // Update indicates an expected call of Update.
-func (mr *MockORMMockRecorder) Update(model interface{}) *gomock.Call {
+func (mr *MockORMMockRecorder) Update(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockORM)(nil).Update), model)
 }
@@ -296,6 +387,7 @@ func (mr *MockORMMockRecorder) Update(model interface{}) *gomock.Call {
 type MockDB struct {
 	ctrl     *gomock.Controller
 	recorder *MockDBMockRecorder
+	isgomock struct{}
 }
 
 // MockDBMockRecorder is the mock recorder for MockDB.
@@ -359,9 +451,9 @@ func (mr *MockDBMockRecorder) Context() *gomock.Call {
 }
 
 // CopyFrom mocks base method.
-func (m *MockDB) CopyFrom(r io.Reader, query interface{}, params ...interface{}) (orm.Result, error) {
+func (m *MockDB) CopyFrom(r io.Reader, query any, params ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{r, query}
+	varargs := []any{r, query}
 	for _, a := range params {
 		varargs = append(varargs, a)
 	}
@@ -372,16 +464,16 @@ func (m *MockDB) CopyFrom(r io.Reader, query interface{}, params ...interface{})
 }
 
 // CopyFrom indicates an expected call of CopyFrom.
-func (mr *MockDBMockRecorder) CopyFrom(r, query interface{}, params ...interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) CopyFrom(r, query any, params ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{r, query}, params...)
+	varargs := append([]any{r, query}, params...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CopyFrom", reflect.TypeOf((*MockDB)(nil).CopyFrom), varargs...)
 }
 
 // CopyTo mocks base method.
-func (m *MockDB) CopyTo(w io.Writer, query interface{}, params ...interface{}) (orm.Result, error) {
+func (m *MockDB) CopyTo(w io.Writer, query any, params ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{w, query}
+	varargs := []any{w, query}
 	for _, a := range params {
 		varargs = append(varargs, a)
 	}
@@ -392,30 +484,31 @@ func (m *MockDB) CopyTo(w io.Writer, query interface{}, params ...interface{}) (
 }
 
 // CopyTo indicates an expected call of CopyTo.
-func (mr *MockDBMockRecorder) CopyTo(w, query interface{}, params ...interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) CopyTo(w, query any, params ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{w, query}, params...)
+	varargs := append([]any{w, query}, params...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CopyTo", reflect.TypeOf((*MockDB)(nil).CopyTo), varargs...)
 }
 
 // Delete mocks base method.
-func (m *MockDB) Delete(model interface{}) error {
+func (m *MockDB) Delete(model any) (orm.Result, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Delete", model)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // Delete indicates an expected call of Delete.
-func (mr *MockDBMockRecorder) Delete(model interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) Delete(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Delete", reflect.TypeOf((*MockDB)(nil).Delete), model)
 }
 
 // Exec mocks base method.
-func (m *MockDB) Exec(arg0 interface{}, arg1 ...interface{}) (orm.Result, error) {
+func (m *MockDB) Exec(arg0 any, arg1 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0}
+	varargs := []any{arg0}
 	for _, a := range arg1 {
 		varargs = append(varargs, a)
 	}
@@ -426,16 +519,36 @@ func (m *MockDB) Exec(arg0 interface{}, arg1 ...interface{}) (orm.Result, error)
 }
 
 // Exec indicates an expected call of Exec.
-func (mr *MockDBMockRecorder) Exec(arg0 interface{}, arg1 ...interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) Exec(arg0 any, arg1 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0}, arg1...)
+	varargs := append([]any{arg0}, arg1...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Exec", reflect.TypeOf((*MockDB)(nil).Exec), varargs...)
 }
 
-// ExecOne mocks base method.
-func (m *MockDB) ExecOne(arg0 interface{}, arg1 ...interface{}) (orm.Result, error) {
+// ExecContext mocks base method.
+func (m *MockDB) ExecContext(arg0 context.Context, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0}
+	varargs := []any{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ExecContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ExecContext indicates an expected call of ExecContext.
+func (mr *MockDBMockRecorder) ExecContext(arg0, arg1 any, arg2 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecContext", reflect.TypeOf((*MockDB)(nil).ExecContext), varargs...)
+}
+
+// ExecOne mocks base method.
+func (m *MockDB) ExecOne(arg0 any, arg1 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0}
 	for _, a := range arg1 {
 		varargs = append(varargs, a)
 	}
@@ -446,30 +559,51 @@ func (m *MockDB) ExecOne(arg0 interface{}, arg1 ...interface{}) (orm.Result, err
 }
 
 // ExecOne indicates an expected call of ExecOne.
-func (mr *MockDBMockRecorder) ExecOne(arg0 interface{}, arg1 ...interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) ExecOne(arg0 any, arg1 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0}, arg1...)
+	varargs := append([]any{arg0}, arg1...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecOne", reflect.TypeOf((*MockDB)(nil).ExecOne), varargs...)
 }
 
+// ExecOneContext mocks base method.
+func (m *MockDB) ExecOneContext(arg0 context.Context, arg1 any, arg2 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ExecOneContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ExecOneContext indicates an expected call of ExecOneContext.
+func (mr *MockDBMockRecorder) ExecOneContext(arg0, arg1 any, arg2 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecOneContext", reflect.TypeOf((*MockDB)(nil).ExecOneContext), varargs...)
+}
+
 // ForceDelete mocks base method.
-func (m *MockDB) ForceDelete(model interface{}) error {
+func (m *MockDB) ForceDelete(model any) (orm.Result, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "ForceDelete", model)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // ForceDelete indicates an expected call of ForceDelete.
-func (mr *MockDBMockRecorder) ForceDelete(model interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) ForceDelete(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ForceDelete", reflect.TypeOf((*MockDB)(nil).ForceDelete), model)
 }
 
 // FormatQuery mocks base method.
-func (m *MockDB) FormatQuery(b []byte, query string, params ...interface{}) []byte {
+func (m *MockDB) FormatQuery(b []byte, query string, params ...any) []byte {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{b, query}
+	varargs := []any{b, query}
 	for _, a := range params {
 		varargs = append(varargs, a)
 	}
@@ -479,34 +613,35 @@ func (m *MockDB) FormatQuery(b []byte, query string, params ...interface{}) []by
 }
 
 // FormatQuery indicates an expected call of FormatQuery.
-func (mr *MockDBMockRecorder) FormatQuery(b, query interface{}, params ...interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) FormatQuery(b, query any, params ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{b, query}, params...)
+	varargs := append([]any{b, query}, params...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "FormatQuery", reflect.TypeOf((*MockDB)(nil).FormatQuery), varargs...)
 }
 
 // Insert mocks base method.
-func (m *MockDB) Insert(model ...interface{}) error {
+func (m *MockDB) Insert(model ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{}
 	for _, a := range model {
 		varargs = append(varargs, a)
 	}
 	ret := m.ctrl.Call(m, "Insert", varargs...)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // Insert indicates an expected call of Insert.
-func (mr *MockDBMockRecorder) Insert(model ...interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) Insert(model ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Insert", reflect.TypeOf((*MockDB)(nil).Insert), model...)
 }
 
 // Model mocks base method.
-func (m *MockDB) Model(model ...interface{}) *orm.Query {
+func (m *MockDB) Model(model ...any) *orm.Query {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{}
 	for _, a := range model {
 		varargs = append(varargs, a)
 	}
@@ -516,15 +651,15 @@ func (m *MockDB) Model(model ...interface{}) *orm.Query {
 }
 
 // Model indicates an expected call of Model.
-func (mr *MockDBMockRecorder) Model(model ...interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) Model(model ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Model", reflect.TypeOf((*MockDB)(nil).Model), model...)
 }
 
 // Query mocks base method.
-func (m *MockDB) Query(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result, error) {
+func (m *MockDB) Query(arg0, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1}
+	varargs := []any{arg0, arg1}
 	for _, a := range arg2 {
 		varargs = append(varargs, a)
 	}
@@ -535,16 +670,36 @@ func (m *MockDB) Query(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result,
 }
 
 // Query indicates an expected call of Query.
-func (mr *MockDBMockRecorder) Query(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) Query(arg0, arg1 any, arg2 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	varargs := append([]any{arg0, arg1}, arg2...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Query", reflect.TypeOf((*MockDB)(nil).Query), varargs...)
 }
 
-// QueryOne mocks base method.
-func (m *MockDB) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result, error) {
+// QueryContext mocks base method.
+func (m *MockDB) QueryContext(arg0 context.Context, arg1, arg2 any, arg3 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1}
+	varargs := []any{arg0, arg1, arg2}
+	for _, a := range arg3 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "QueryContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// QueryContext indicates an expected call of QueryContext.
+func (mr *MockDBMockRecorder) QueryContext(arg0, arg1, arg2 any, arg3 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1, arg2}, arg3...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryContext", reflect.TypeOf((*MockDB)(nil).QueryContext), varargs...)
+}
+
+// QueryOne mocks base method.
+func (m *MockDB) QueryOne(arg0, arg1 any, arg2 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1}
 	for _, a := range arg2 {
 		varargs = append(varargs, a)
 	}
@@ -555,14 +710,34 @@ func (m *MockDB) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Resu
 }
 
 // QueryOne indicates an expected call of QueryOne.
-func (mr *MockDBMockRecorder) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) QueryOne(arg0, arg1 any, arg2 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	varargs := append([]any{arg0, arg1}, arg2...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryOne", reflect.TypeOf((*MockDB)(nil).QueryOne), varargs...)
 }
 
+// QueryOneContext mocks base method.
+func (m *MockDB) QueryOneContext(arg0 context.Context, arg1, arg2 any, arg3 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1, arg2}
+	for _, a := range arg3 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "QueryOneContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// QueryOneContext indicates an expected call of QueryOneContext.
+func (mr *MockDBMockRecorder) QueryOneContext(arg0, arg1, arg2 any, arg3 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1, arg2}, arg3...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryOneContext", reflect.TypeOf((*MockDB)(nil).QueryOneContext), varargs...)
+}
+
 // Select mocks base method.
-func (m *MockDB) Select(model interface{}) error {
+func (m *MockDB) Select(model any) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Select", model)
 	ret0, _ := ret[0].(error)
@@ -570,21 +745,22 @@ func (m *MockDB) Select(model interface{}) error {
 }
 
 // Select indicates an expected call of Select.
-func (mr *MockDBMockRecorder) Select(model interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) Select(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Select", reflect.TypeOf((*MockDB)(nil).Select), model)
 }
 
 // Update mocks base method.
-func (m *MockDB) Update(model interface{}) error {
+func (m *MockDB) Update(model any) (orm.Result, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Update", model)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // Update indicates an expected call of Update.
-func (mr *MockDBMockRecorder) Update(model interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) Update(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockDB)(nil).Update), model)
 }
@@ -598,7 +774,7 @@ func (m *MockDB) WithContext(ctx context.Context) *pg.DB {
 }
 
 // WithContext indicates an expected call of WithContext.
-func (mr *MockDBMockRecorder) WithContext(ctx interface{}) *gomock.Call {
+func (mr *MockDBMockRecorder) WithContext(ctx any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "WithContext", reflect.TypeOf((*MockDB)(nil).WithContext), ctx)
 }
@@ -607,6 +783,7 @@ func (mr *MockDBMockRecorder) WithContext(ctx interface{}) *gomock.Call {
 type MockTx struct {
 	ctrl     *gomock.Controller
 	recorder *MockTxMockRecorder
+	isgomock struct{}
 }
 
 // MockTxMockRecorder is the mock recorder for MockTx.
@@ -641,9 +818,9 @@ func (mr *MockTxMockRecorder) Commit() *gomock.Call {
 }
 
 // Exec mocks base method.
-func (m *MockTx) Exec(arg0 interface{}, arg1 ...interface{}) (orm.Result, error) {
+func (m *MockTx) Exec(arg0 any, arg1 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0}
+	varargs := []any{arg0}
 	for _, a := range arg1 {
 		varargs = append(varargs, a)
 	}
@@ -654,16 +831,36 @@ func (m *MockTx) Exec(arg0 interface{}, arg1 ...interface{}) (orm.Result, error)
 }
 
 // Exec indicates an expected call of Exec.
-func (mr *MockTxMockRecorder) Exec(arg0 interface{}, arg1 ...interface{}) *gomock.Call {
+func (mr *MockTxMockRecorder) Exec(arg0 any, arg1 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0}, arg1...)
+	varargs := append([]any{arg0}, arg1...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Exec", reflect.TypeOf((*MockTx)(nil).Exec), varargs...)
 }
 
-// ExecOne mocks base method.
-func (m *MockTx) ExecOne(arg0 interface{}, arg1 ...interface{}) (orm.Result, error) {
+// ExecContext mocks base method.
+func (m *MockTx) ExecContext(arg0 context.Context, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0}
+	varargs := []any{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ExecContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ExecContext indicates an expected call of ExecContext.
+func (mr *MockTxMockRecorder) ExecContext(arg0, arg1 any, arg2 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecContext", reflect.TypeOf((*MockTx)(nil).ExecContext), varargs...)
+}
+
+// ExecOne mocks base method.
+func (m *MockTx) ExecOne(arg0 any, arg1 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0}
 	for _, a := range arg1 {
 		varargs = append(varargs, a)
 	}
@@ -674,16 +871,36 @@ func (m *MockTx) ExecOne(arg0 interface{}, arg1 ...interface{}) (orm.Result, err
 }
 
 // ExecOne indicates an expected call of ExecOne.
-func (mr *MockTxMockRecorder) ExecOne(arg0 interface{}, arg1 ...interface{}) *gomock.Call {
+func (mr *MockTxMockRecorder) ExecOne(arg0 any, arg1 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0}, arg1...)
+	varargs := append([]any{arg0}, arg1...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecOne", reflect.TypeOf((*MockTx)(nil).ExecOne), varargs...)
 }
 
-// Model mocks base method.
-func (m *MockTx) Model(model ...interface{}) *orm.Query {
+// ExecOneContext mocks base method.
+func (m *MockTx) ExecOneContext(arg0 context.Context, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ExecOneContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ExecOneContext indicates an expected call of ExecOneContext.
+func (mr *MockTxMockRecorder) ExecOneContext(arg0, arg1 any, arg2 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecOneContext", reflect.TypeOf((*MockTx)(nil).ExecOneContext), varargs...)
+}
+
+// Model mocks base method.
+func (m *MockTx) Model(model ...any) *orm.Query {
+	m.ctrl.T.Helper()
+	varargs := []any{}
 	for _, a := range model {
 		varargs = append(varargs, a)
 	}
@@ -693,15 +910,15 @@ func (m *MockTx) Model(model ...interface{}) *orm.Query {
 }
 
 // Model indicates an expected call of Model.
-func (mr *MockTxMockRecorder) Model(model ...interface{}) *gomock.Call {
+func (mr *MockTxMockRecorder) Model(model ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Model", reflect.TypeOf((*MockTx)(nil).Model), model...)
 }
 
 // Query mocks base method.
-func (m *MockTx) Query(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result, error) {
+func (m *MockTx) Query(arg0, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1}
+	varargs := []any{arg0, arg1}
 	for _, a := range arg2 {
 		varargs = append(varargs, a)
 	}
@@ -712,16 +929,36 @@ func (m *MockTx) Query(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result,
 }
 
 // Query indicates an expected call of Query.
-func (mr *MockTxMockRecorder) Query(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+func (mr *MockTxMockRecorder) Query(arg0, arg1 any, arg2 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	varargs := append([]any{arg0, arg1}, arg2...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Query", reflect.TypeOf((*MockTx)(nil).Query), varargs...)
 }
 
-// QueryOne mocks base method.
-func (m *MockTx) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result, error) {
+// QueryContext mocks base method.
+func (m *MockTx) QueryContext(arg0 context.Context, arg1, arg2 any, arg3 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1}
+	varargs := []any{arg0, arg1, arg2}
+	for _, a := range arg3 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "QueryContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// QueryContext indicates an expected call of QueryContext.
+func (mr *MockTxMockRecorder) QueryContext(arg0, arg1, arg2 any, arg3 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1, arg2}, arg3...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryContext", reflect.TypeOf((*MockTx)(nil).QueryContext), varargs...)
+}
+
+// QueryOne mocks base method.
+func (m *MockTx) QueryOne(arg0, arg1 any, arg2 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1}
 	for _, a := range arg2 {
 		varargs = append(varargs, a)
 	}
@@ -732,10 +969,30 @@ func (m *MockTx) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Resu
 }
 
 // QueryOne indicates an expected call of QueryOne.
-func (mr *MockTxMockRecorder) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+func (mr *MockTxMockRecorder) QueryOne(arg0, arg1 any, arg2 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	varargs := append([]any{arg0, arg1}, arg2...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryOne", reflect.TypeOf((*MockTx)(nil).QueryOne), varargs...)
+}
+
+// QueryOneContext mocks base method.
+func (m *MockTx) QueryOneContext(arg0 context.Context, arg1, arg2 any, arg3 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1, arg2}
+	for _, a := range arg3 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "QueryOneContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// QueryOneContext indicates an expected call of QueryOneContext.
+func (mr *MockTxMockRecorder) QueryOneContext(arg0, arg1, arg2 any, arg3 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1, arg2}, arg3...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryOneContext", reflect.TypeOf((*MockTx)(nil).QueryOneContext), varargs...)
 }
 
 // Rollback mocks base method.
@@ -756,6 +1013,7 @@ func (mr *MockTxMockRecorder) Rollback() *gomock.Call {
 type MockTxWrapper struct {
 	ctrl     *gomock.Controller
 	recorder *MockTxWrapperMockRecorder
+	isgomock struct{}
 }
 
 // MockTxWrapperMockRecorder is the mock recorder for MockTxWrapper.
@@ -785,7 +1043,7 @@ func (m *MockTxWrapper) DbBegin(db interfaces.DB) (interfaces.DB, error) {
 }
 
 // DbBegin indicates an expected call of DbBegin.
-func (mr *MockTxWrapperMockRecorder) DbBegin(db interface{}) *gomock.Call {
+func (mr *MockTxWrapperMockRecorder) DbBegin(db any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DbBegin", reflect.TypeOf((*MockTxWrapper)(nil).DbBegin), db)
 }
@@ -794,6 +1052,7 @@ func (mr *MockTxWrapperMockRecorder) DbBegin(db interface{}) *gomock.Call {
 type MockCtxWrapper struct {
 	ctrl     *gomock.Controller
 	recorder *MockCtxWrapperMockRecorder
+	isgomock struct{}
 }
 
 // MockCtxWrapperMockRecorder is the mock recorder for MockCtxWrapper.
@@ -822,7 +1081,7 @@ func (m *MockCtxWrapper) WithContext(ctx context.Context, db interfaces.DB) inte
 }
 
 // WithContext indicates an expected call of WithContext.
-func (mr *MockCtxWrapperMockRecorder) WithContext(ctx, db interface{}) *gomock.Call {
+func (mr *MockCtxWrapperMockRecorder) WithContext(ctx, db any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "WithContext", reflect.TypeOf((*MockCtxWrapper)(nil).WithContext), ctx, db)
 }
@@ -831,6 +1090,7 @@ func (mr *MockCtxWrapperMockRecorder) WithContext(ctx, db interface{}) *gomock.C
 type MockDBTx struct {
 	ctrl     *gomock.Controller
 	recorder *MockDBTxMockRecorder
+	isgomock struct{}
 }
 
 // MockDBTxMockRecorder is the mock recorder for MockDBTx.
@@ -908,9 +1168,9 @@ func (mr *MockDBTxMockRecorder) Context() *gomock.Call {
 }
 
 // CopyFrom mocks base method.
-func (m *MockDBTx) CopyFrom(r io.Reader, query interface{}, params ...interface{}) (orm.Result, error) {
+func (m *MockDBTx) CopyFrom(r io.Reader, query any, params ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{r, query}
+	varargs := []any{r, query}
 	for _, a := range params {
 		varargs = append(varargs, a)
 	}
@@ -921,16 +1181,16 @@ func (m *MockDBTx) CopyFrom(r io.Reader, query interface{}, params ...interface{
 }
 
 // CopyFrom indicates an expected call of CopyFrom.
-func (mr *MockDBTxMockRecorder) CopyFrom(r, query interface{}, params ...interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) CopyFrom(r, query any, params ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{r, query}, params...)
+	varargs := append([]any{r, query}, params...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CopyFrom", reflect.TypeOf((*MockDBTx)(nil).CopyFrom), varargs...)
 }
 
 // CopyTo mocks base method.
-func (m *MockDBTx) CopyTo(w io.Writer, query interface{}, params ...interface{}) (orm.Result, error) {
+func (m *MockDBTx) CopyTo(w io.Writer, query any, params ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{w, query}
+	varargs := []any{w, query}
 	for _, a := range params {
 		varargs = append(varargs, a)
 	}
@@ -941,30 +1201,31 @@ func (m *MockDBTx) CopyTo(w io.Writer, query interface{}, params ...interface{})
 }
 
 // CopyTo indicates an expected call of CopyTo.
-func (mr *MockDBTxMockRecorder) CopyTo(w, query interface{}, params ...interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) CopyTo(w, query any, params ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{w, query}, params...)
+	varargs := append([]any{w, query}, params...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "CopyTo", reflect.TypeOf((*MockDBTx)(nil).CopyTo), varargs...)
 }
 
 // Delete mocks base method.
-func (m *MockDBTx) Delete(model interface{}) error {
+func (m *MockDBTx) Delete(model any) (orm.Result, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Delete", model)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // Delete indicates an expected call of Delete.
-func (mr *MockDBTxMockRecorder) Delete(model interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) Delete(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Delete", reflect.TypeOf((*MockDBTx)(nil).Delete), model)
 }
 
 // Exec mocks base method.
-func (m *MockDBTx) Exec(arg0 interface{}, arg1 ...interface{}) (orm.Result, error) {
+func (m *MockDBTx) Exec(arg0 any, arg1 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0}
+	varargs := []any{arg0}
 	for _, a := range arg1 {
 		varargs = append(varargs, a)
 	}
@@ -975,16 +1236,36 @@ func (m *MockDBTx) Exec(arg0 interface{}, arg1 ...interface{}) (orm.Result, erro
 }
 
 // Exec indicates an expected call of Exec.
-func (mr *MockDBTxMockRecorder) Exec(arg0 interface{}, arg1 ...interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) Exec(arg0 any, arg1 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0}, arg1...)
+	varargs := append([]any{arg0}, arg1...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Exec", reflect.TypeOf((*MockDBTx)(nil).Exec), varargs...)
 }
 
-// ExecOne mocks base method.
-func (m *MockDBTx) ExecOne(arg0 interface{}, arg1 ...interface{}) (orm.Result, error) {
+// ExecContext mocks base method.
+func (m *MockDBTx) ExecContext(arg0 context.Context, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0}
+	varargs := []any{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ExecContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ExecContext indicates an expected call of ExecContext.
+func (mr *MockDBTxMockRecorder) ExecContext(arg0, arg1 any, arg2 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecContext", reflect.TypeOf((*MockDBTx)(nil).ExecContext), varargs...)
+}
+
+// ExecOne mocks base method.
+func (m *MockDBTx) ExecOne(arg0 any, arg1 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0}
 	for _, a := range arg1 {
 		varargs = append(varargs, a)
 	}
@@ -995,30 +1276,51 @@ func (m *MockDBTx) ExecOne(arg0 interface{}, arg1 ...interface{}) (orm.Result, e
 }
 
 // ExecOne indicates an expected call of ExecOne.
-func (mr *MockDBTxMockRecorder) ExecOne(arg0 interface{}, arg1 ...interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) ExecOne(arg0 any, arg1 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0}, arg1...)
+	varargs := append([]any{arg0}, arg1...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecOne", reflect.TypeOf((*MockDBTx)(nil).ExecOne), varargs...)
 }
 
+// ExecOneContext mocks base method.
+func (m *MockDBTx) ExecOneContext(arg0 context.Context, arg1 any, arg2 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1}
+	for _, a := range arg2 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "ExecOneContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// ExecOneContext indicates an expected call of ExecOneContext.
+func (mr *MockDBTxMockRecorder) ExecOneContext(arg0, arg1 any, arg2 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1}, arg2...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ExecOneContext", reflect.TypeOf((*MockDBTx)(nil).ExecOneContext), varargs...)
+}
+
 // ForceDelete mocks base method.
-func (m *MockDBTx) ForceDelete(model interface{}) error {
+func (m *MockDBTx) ForceDelete(model any) (orm.Result, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "ForceDelete", model)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // ForceDelete indicates an expected call of ForceDelete.
-func (mr *MockDBTxMockRecorder) ForceDelete(model interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) ForceDelete(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ForceDelete", reflect.TypeOf((*MockDBTx)(nil).ForceDelete), model)
 }
 
 // FormatQuery mocks base method.
-func (m *MockDBTx) FormatQuery(b []byte, query string, params ...interface{}) []byte {
+func (m *MockDBTx) FormatQuery(b []byte, query string, params ...any) []byte {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{b, query}
+	varargs := []any{b, query}
 	for _, a := range params {
 		varargs = append(varargs, a)
 	}
@@ -1028,34 +1330,35 @@ func (m *MockDBTx) FormatQuery(b []byte, query string, params ...interface{}) []
 }
 
 // FormatQuery indicates an expected call of FormatQuery.
-func (mr *MockDBTxMockRecorder) FormatQuery(b, query interface{}, params ...interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) FormatQuery(b, query any, params ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{b, query}, params...)
+	varargs := append([]any{b, query}, params...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "FormatQuery", reflect.TypeOf((*MockDBTx)(nil).FormatQuery), varargs...)
 }
 
 // Insert mocks base method.
-func (m *MockDBTx) Insert(model ...interface{}) error {
+func (m *MockDBTx) Insert(model ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{}
 	for _, a := range model {
 		varargs = append(varargs, a)
 	}
 	ret := m.ctrl.Call(m, "Insert", varargs...)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // Insert indicates an expected call of Insert.
-func (mr *MockDBTxMockRecorder) Insert(model ...interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) Insert(model ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Insert", reflect.TypeOf((*MockDBTx)(nil).Insert), model...)
 }
 
 // Model mocks base method.
-func (m *MockDBTx) Model(model ...interface{}) *orm.Query {
+func (m *MockDBTx) Model(model ...any) *orm.Query {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{}
 	for _, a := range model {
 		varargs = append(varargs, a)
 	}
@@ -1065,15 +1368,15 @@ func (m *MockDBTx) Model(model ...interface{}) *orm.Query {
 }
 
 // Model indicates an expected call of Model.
-func (mr *MockDBTxMockRecorder) Model(model ...interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) Model(model ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Model", reflect.TypeOf((*MockDBTx)(nil).Model), model...)
 }
 
 // Query mocks base method.
-func (m *MockDBTx) Query(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result, error) {
+func (m *MockDBTx) Query(arg0, arg1 any, arg2 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1}
+	varargs := []any{arg0, arg1}
 	for _, a := range arg2 {
 		varargs = append(varargs, a)
 	}
@@ -1084,16 +1387,36 @@ func (m *MockDBTx) Query(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Resul
 }
 
 // Query indicates an expected call of Query.
-func (mr *MockDBTxMockRecorder) Query(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) Query(arg0, arg1 any, arg2 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	varargs := append([]any{arg0, arg1}, arg2...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Query", reflect.TypeOf((*MockDBTx)(nil).Query), varargs...)
 }
 
-// QueryOne mocks base method.
-func (m *MockDBTx) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Result, error) {
+// QueryContext mocks base method.
+func (m *MockDBTx) QueryContext(arg0 context.Context, arg1, arg2 any, arg3 ...any) (orm.Result, error) {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0, arg1}
+	varargs := []any{arg0, arg1, arg2}
+	for _, a := range arg3 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "QueryContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// QueryContext indicates an expected call of QueryContext.
+func (mr *MockDBTxMockRecorder) QueryContext(arg0, arg1, arg2 any, arg3 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1, arg2}, arg3...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryContext", reflect.TypeOf((*MockDBTx)(nil).QueryContext), varargs...)
+}
+
+// QueryOne mocks base method.
+func (m *MockDBTx) QueryOne(arg0, arg1 any, arg2 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1}
 	for _, a := range arg2 {
 		varargs = append(varargs, a)
 	}
@@ -1104,10 +1427,30 @@ func (m *MockDBTx) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) (orm.Re
 }
 
 // QueryOne indicates an expected call of QueryOne.
-func (mr *MockDBTxMockRecorder) QueryOne(arg0, arg1 interface{}, arg2 ...interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) QueryOne(arg0, arg1 any, arg2 ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0, arg1}, arg2...)
+	varargs := append([]any{arg0, arg1}, arg2...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryOne", reflect.TypeOf((*MockDBTx)(nil).QueryOne), varargs...)
+}
+
+// QueryOneContext mocks base method.
+func (m *MockDBTx) QueryOneContext(arg0 context.Context, arg1, arg2 any, arg3 ...any) (orm.Result, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{arg0, arg1, arg2}
+	for _, a := range arg3 {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "QueryOneContext", varargs...)
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// QueryOneContext indicates an expected call of QueryOneContext.
+func (mr *MockDBTxMockRecorder) QueryOneContext(arg0, arg1, arg2 any, arg3 ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{arg0, arg1, arg2}, arg3...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "QueryOneContext", reflect.TypeOf((*MockDBTx)(nil).QueryOneContext), varargs...)
 }
 
 // Rollback mocks base method.
@@ -1125,7 +1468,7 @@ func (mr *MockDBTxMockRecorder) Rollback() *gomock.Call {
 }
 
 // Select mocks base method.
-func (m *MockDBTx) Select(model interface{}) error {
+func (m *MockDBTx) Select(model any) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Select", model)
 	ret0, _ := ret[0].(error)
@@ -1133,21 +1476,22 @@ func (m *MockDBTx) Select(model interface{}) error {
 }
 
 // Select indicates an expected call of Select.
-func (mr *MockDBTxMockRecorder) Select(model interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) Select(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Select", reflect.TypeOf((*MockDBTx)(nil).Select), model)
 }
 
 // Update mocks base method.
-func (m *MockDBTx) Update(model interface{}) error {
+func (m *MockDBTx) Update(model any) (orm.Result, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "Update", model)
-	ret0, _ := ret[0].(error)
-	return ret0
+	ret0, _ := ret[0].(orm.Result)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
 }
 
 // Update indicates an expected call of Update.
-func (mr *MockDBTxMockRecorder) Update(model interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) Update(model any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Update", reflect.TypeOf((*MockDBTx)(nil).Update), model)
 }
@@ -1161,7 +1505,7 @@ func (m *MockDBTx) WithContext(ctx context.Context) *pg.DB {
 }
 
 // WithContext indicates an expected call of WithContext.
-func (mr *MockDBTxMockRecorder) WithContext(ctx interface{}) *gomock.Call {
+func (mr *MockDBTxMockRecorder) WithContext(ctx any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "WithContext", reflect.TypeOf((*MockDBTx)(nil).WithContext), ctx)
 }
