@@ -1,5 +1,10 @@
 // Code generated by MockGen. DO NOT EDIT.
 // Source: redis/interfaces/redis.go
+//
+// Generated by this command:
+//
+//	mockgen -source=redis/interfaces/redis.go -destination=redis/mocks/redis.go -package mocks
+//
 
 // Package mocks is a generated GoMock package.
 package mocks
@@ -9,15 +14,16 @@ import (
 	reflect "reflect"
 	time "time"
 
-	redis "github.com/go-redis/redis"
-	gomock "github.com/golang/mock/gomock"
+	redis "github.com/redis/go-redis/v9"
 	interfaces "github.com/topfreegames/extensions/v9/redis/interfaces"
+	gomock "go.uber.org/mock/gomock"
 )
 
 // MockTraceWrapper is a mock of TraceWrapper interface.
 type MockTraceWrapper struct {
 	ctrl     *gomock.Controller
 	recorder *MockTraceWrapperMockRecorder
+	isgomock struct{}
 }
 
 // MockTraceWrapperMockRecorder is the mock recorder for MockTraceWrapper.
@@ -46,7 +52,7 @@ func (m *MockTraceWrapper) WithContext(ctx context.Context, c interfaces.RedisCl
 }
 
 // WithContext indicates an expected call of WithContext.
-func (mr *MockTraceWrapperMockRecorder) WithContext(ctx, c interface{}) *gomock.Call {
+func (mr *MockTraceWrapperMockRecorder) WithContext(ctx, c any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "WithContext", reflect.TypeOf((*MockTraceWrapper)(nil).WithContext), ctx, c)
 }
@@ -55,6 +61,7 @@ func (mr *MockTraceWrapperMockRecorder) WithContext(ctx, c interface{}) *gomock.
 type MockRedisClient struct {
 	ctrl     *gomock.Controller
 	recorder *MockRedisClientMockRecorder
+	isgomock struct{}
 }
 
 // MockRedisClientMockRecorder is the mock recorder for MockRedisClient.
@@ -75,9 +82,9 @@ func (m *MockRedisClient) EXPECT() *MockRedisClientMockRecorder {
 }
 
 // BLPop mocks base method.
-func (m *MockRedisClient) BLPop(timeout time.Duration, keys ...string) *redis.StringSliceCmd {
+func (m *MockRedisClient) BLPop(ctx context.Context, timeout time.Duration, keys ...string) *redis.StringSliceCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{timeout}
+	varargs := []any{ctx, timeout}
 	for _, a := range keys {
 		varargs = append(varargs, a)
 	}
@@ -87,9 +94,9 @@ func (m *MockRedisClient) BLPop(timeout time.Duration, keys ...string) *redis.St
 }
 
 // BLPop indicates an expected call of BLPop.
-func (mr *MockRedisClientMockRecorder) BLPop(timeout interface{}, keys ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) BLPop(ctx, timeout any, keys ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{timeout}, keys...)
+	varargs := append([]any{ctx, timeout}, keys...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "BLPop", reflect.TypeOf((*MockRedisClient)(nil).BLPop), varargs...)
 }
 
@@ -107,24 +114,10 @@ func (mr *MockRedisClientMockRecorder) Close() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Close", reflect.TypeOf((*MockRedisClient)(nil).Close))
 }
 
-// Context mocks base method.
-func (m *MockRedisClient) Context() context.Context {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Context")
-	ret0, _ := ret[0].(context.Context)
-	return ret0
-}
-
-// Context indicates an expected call of Context.
-func (mr *MockRedisClientMockRecorder) Context() *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Context", reflect.TypeOf((*MockRedisClient)(nil).Context))
-}
-
 // Del mocks base method.
-func (m *MockRedisClient) Del(keys ...string) *redis.IntCmd {
+func (m *MockRedisClient) Del(ctx context.Context, keys ...string) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{ctx}
 	for _, a := range keys {
 		varargs = append(varargs, a)
 	}
@@ -134,15 +127,16 @@ func (m *MockRedisClient) Del(keys ...string) *redis.IntCmd {
 }
 
 // Del indicates an expected call of Del.
-func (mr *MockRedisClientMockRecorder) Del(keys ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) Del(ctx any, keys ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Del", reflect.TypeOf((*MockRedisClient)(nil).Del), keys...)
+	varargs := append([]any{ctx}, keys...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Del", reflect.TypeOf((*MockRedisClient)(nil).Del), varargs...)
 }
 
 // Eval mocks base method.
-func (m *MockRedisClient) Eval(script string, keys []string, args ...interface{}) *redis.Cmd {
+func (m *MockRedisClient) Eval(ctx context.Context, script string, keys []string, args ...any) *redis.Cmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{script, keys}
+	varargs := []any{ctx, script, keys}
 	for _, a := range args {
 		varargs = append(varargs, a)
 	}
@@ -152,16 +146,16 @@ func (m *MockRedisClient) Eval(script string, keys []string, args ...interface{}
 }
 
 // Eval indicates an expected call of Eval.
-func (mr *MockRedisClientMockRecorder) Eval(script, keys interface{}, args ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) Eval(ctx, script, keys any, args ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{script, keys}, args...)
+	varargs := append([]any{ctx, script, keys}, args...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Eval", reflect.TypeOf((*MockRedisClient)(nil).Eval), varargs...)
 }
 
 // EvalSha mocks base method.
-func (m *MockRedisClient) EvalSha(sha1 string, keys []string, args ...interface{}) *redis.Cmd {
+func (m *MockRedisClient) EvalSha(ctx context.Context, sha1 string, keys []string, args ...any) *redis.Cmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{sha1, keys}
+	varargs := []any{ctx, sha1, keys}
 	for _, a := range args {
 		varargs = append(varargs, a)
 	}
@@ -171,16 +165,16 @@ func (m *MockRedisClient) EvalSha(sha1 string, keys []string, args ...interface{
 }
 
 // EvalSha indicates an expected call of EvalSha.
-func (mr *MockRedisClientMockRecorder) EvalSha(sha1, keys interface{}, args ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) EvalSha(ctx, sha1, keys any, args ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{sha1, keys}, args...)
+	varargs := append([]any{ctx, sha1, keys}, args...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "EvalSha", reflect.TypeOf((*MockRedisClient)(nil).EvalSha), varargs...)
 }
 
 // Exists mocks base method.
-func (m *MockRedisClient) Exists(keys ...string) *redis.IntCmd {
+func (m *MockRedisClient) Exists(ctx context.Context, keys ...string) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{ctx}
 	for _, a := range keys {
 		varargs = append(varargs, a)
 	}
@@ -190,29 +184,30 @@ func (m *MockRedisClient) Exists(keys ...string) *redis.IntCmd {
 }
 
 // Exists indicates an expected call of Exists.
-func (mr *MockRedisClientMockRecorder) Exists(keys ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) Exists(ctx any, keys ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Exists", reflect.TypeOf((*MockRedisClient)(nil).Exists), keys...)
+	varargs := append([]any{ctx}, keys...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Exists", reflect.TypeOf((*MockRedisClient)(nil).Exists), varargs...)
 }
 
 // Get mocks base method.
-func (m *MockRedisClient) Get(key string) *redis.StringCmd {
+func (m *MockRedisClient) Get(ctx context.Context, key string) *redis.StringCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Get", key)
+	ret := m.ctrl.Call(m, "Get", ctx, key)
 	ret0, _ := ret[0].(*redis.StringCmd)
 	return ret0
 }
 
 // Get indicates an expected call of Get.
-func (mr *MockRedisClientMockRecorder) Get(key interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) Get(ctx, key any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Get", reflect.TypeOf((*MockRedisClient)(nil).Get), key)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Get", reflect.TypeOf((*MockRedisClient)(nil).Get), ctx, key)
 }
 
 // HDel mocks base method.
-func (m *MockRedisClient) HDel(key string, fields ...string) *redis.IntCmd {
+func (m *MockRedisClient) HDel(ctx context.Context, key string, fields ...string) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{key}
+	varargs := []any{ctx, key}
 	for _, a := range fields {
 		varargs = append(varargs, a)
 	}
@@ -222,45 +217,45 @@ func (m *MockRedisClient) HDel(key string, fields ...string) *redis.IntCmd {
 }
 
 // HDel indicates an expected call of HDel.
-func (mr *MockRedisClientMockRecorder) HDel(key interface{}, fields ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) HDel(ctx, key any, fields ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{key}, fields...)
+	varargs := append([]any{ctx, key}, fields...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HDel", reflect.TypeOf((*MockRedisClient)(nil).HDel), varargs...)
 }
 
 // HGet mocks base method.
-func (m *MockRedisClient) HGet(key, field string) *redis.StringCmd {
+func (m *MockRedisClient) HGet(ctx context.Context, key, field string) *redis.StringCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "HGet", key, field)
+	ret := m.ctrl.Call(m, "HGet", ctx, key, field)
 	ret0, _ := ret[0].(*redis.StringCmd)
 	return ret0
 }
 
 // HGet indicates an expected call of HGet.
-func (mr *MockRedisClientMockRecorder) HGet(key, field interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) HGet(ctx, key, field any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HGet", reflect.TypeOf((*MockRedisClient)(nil).HGet), key, field)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HGet", reflect.TypeOf((*MockRedisClient)(nil).HGet), ctx, key, field)
 }
 
 // HGetAll mocks base method.
-func (m *MockRedisClient) HGetAll(arg0 string) *redis.StringStringMapCmd {
+func (m *MockRedisClient) HGetAll(ctx context.Context, key string) *redis.MapStringStringCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "HGetAll", arg0)
-	ret0, _ := ret[0].(*redis.StringStringMapCmd)
+	ret := m.ctrl.Call(m, "HGetAll", ctx, key)
+	ret0, _ := ret[0].(*redis.MapStringStringCmd)
 	return ret0
 }
 
 // HGetAll indicates an expected call of HGetAll.
-func (mr *MockRedisClientMockRecorder) HGetAll(arg0 interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) HGetAll(ctx, key any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HGetAll", reflect.TypeOf((*MockRedisClient)(nil).HGetAll), arg0)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HGetAll", reflect.TypeOf((*MockRedisClient)(nil).HGetAll), ctx, key)
 }
 
 // HMGet mocks base method.
-func (m *MockRedisClient) HMGet(arg0 string, arg1 ...string) *redis.SliceCmd {
+func (m *MockRedisClient) HMGet(ctx context.Context, key string, fields ...string) *redis.SliceCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{arg0}
-	for _, a := range arg1 {
+	varargs := []any{ctx, key}
+	for _, a := range fields {
 		varargs = append(varargs, a)
 	}
 	ret := m.ctrl.Call(m, "HMGet", varargs...)
@@ -269,58 +264,68 @@ func (m *MockRedisClient) HMGet(arg0 string, arg1 ...string) *redis.SliceCmd {
 }
 
 // HMGet indicates an expected call of HMGet.
-func (mr *MockRedisClientMockRecorder) HMGet(arg0 interface{}, arg1 ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) HMGet(ctx, key any, fields ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{arg0}, arg1...)
+	varargs := append([]any{ctx, key}, fields...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HMGet", reflect.TypeOf((*MockRedisClient)(nil).HMGet), varargs...)
 }
 
 // HMSet mocks base method.
-func (m *MockRedisClient) HMSet(arg0 string, arg1 map[string]interface{}) *redis.StatusCmd {
+func (m *MockRedisClient) HMSet(ctx context.Context, key string, values ...any) *redis.BoolCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "HMSet", arg0, arg1)
-	ret0, _ := ret[0].(*redis.StatusCmd)
-	return ret0
-}
-
-// HMSet indicates an expected call of HMSet.
-func (mr *MockRedisClientMockRecorder) HMSet(arg0, arg1 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HMSet", reflect.TypeOf((*MockRedisClient)(nil).HMSet), arg0, arg1)
-}
-
-// HSet mocks base method.
-func (m *MockRedisClient) HSet(key, field string, value interface{}) *redis.BoolCmd {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "HSet", key, field, value)
+	varargs := []any{ctx, key}
+	for _, a := range values {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "HMSet", varargs...)
 	ret0, _ := ret[0].(*redis.BoolCmd)
 	return ret0
 }
 
-// HSet indicates an expected call of HSet.
-func (mr *MockRedisClientMockRecorder) HSet(key, field, value interface{}) *gomock.Call {
+// HMSet indicates an expected call of HMSet.
+func (mr *MockRedisClientMockRecorder) HMSet(ctx, key any, values ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HSet", reflect.TypeOf((*MockRedisClient)(nil).HSet), key, field, value)
+	varargs := append([]any{ctx, key}, values...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HMSet", reflect.TypeOf((*MockRedisClient)(nil).HMSet), varargs...)
+}
+
+// HSet mocks base method.
+func (m *MockRedisClient) HSet(ctx context.Context, key string, values ...any) *redis.IntCmd {
+	m.ctrl.T.Helper()
+	varargs := []any{ctx, key}
+	for _, a := range values {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "HSet", varargs...)
+	ret0, _ := ret[0].(*redis.IntCmd)
+	return ret0
+}
+
+// HSet indicates an expected call of HSet.
+func (mr *MockRedisClientMockRecorder) HSet(ctx, key any, values ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{ctx, key}, values...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HSet", reflect.TypeOf((*MockRedisClient)(nil).HSet), varargs...)
 }
 
 // LRange mocks base method.
-func (m *MockRedisClient) LRange(key string, start, stop int64) *redis.StringSliceCmd {
+func (m *MockRedisClient) LRange(ctx context.Context, key string, start, stop int64) *redis.StringSliceCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "LRange", key, start, stop)
+	ret := m.ctrl.Call(m, "LRange", ctx, key, start, stop)
 	ret0, _ := ret[0].(*redis.StringSliceCmd)
 	return ret0
 }
 
 // LRange indicates an expected call of LRange.
-func (mr *MockRedisClientMockRecorder) LRange(key, start, stop interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) LRange(ctx, key, start, stop any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "LRange", reflect.TypeOf((*MockRedisClient)(nil).LRange), key, start, stop)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "LRange", reflect.TypeOf((*MockRedisClient)(nil).LRange), ctx, key, start, stop)
 }
 
 // MGet mocks base method.
-func (m *MockRedisClient) MGet(keys ...string) *redis.SliceCmd {
+func (m *MockRedisClient) MGet(ctx context.Context, keys ...string) *redis.SliceCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{ctx}
 	for _, a := range keys {
 		varargs = append(varargs, a)
 	}
@@ -330,43 +335,44 @@ func (m *MockRedisClient) MGet(keys ...string) *redis.SliceCmd {
 }
 
 // MGet indicates an expected call of MGet.
-func (mr *MockRedisClientMockRecorder) MGet(keys ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) MGet(ctx any, keys ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "MGet", reflect.TypeOf((*MockRedisClient)(nil).MGet), keys...)
+	varargs := append([]any{ctx}, keys...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "MGet", reflect.TypeOf((*MockRedisClient)(nil).MGet), varargs...)
 }
 
 // Ping mocks base method.
-func (m *MockRedisClient) Ping() *redis.StatusCmd {
+func (m *MockRedisClient) Ping(ctx context.Context) *redis.StatusCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Ping")
+	ret := m.ctrl.Call(m, "Ping", ctx)
 	ret0, _ := ret[0].(*redis.StatusCmd)
 	return ret0
 }
 
 // Ping indicates an expected call of Ping.
-func (mr *MockRedisClientMockRecorder) Ping() *gomock.Call {
+func (mr *MockRedisClientMockRecorder) Ping(ctx any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Ping", reflect.TypeOf((*MockRedisClient)(nil).Ping))
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Ping", reflect.TypeOf((*MockRedisClient)(nil).Ping), ctx)
 }
 
 // RPopLPush mocks base method.
-func (m *MockRedisClient) RPopLPush(source, destination string) *redis.StringCmd {
+func (m *MockRedisClient) RPopLPush(ctx context.Context, source, destination string) *redis.StringCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "RPopLPush", source, destination)
+	ret := m.ctrl.Call(m, "RPopLPush", ctx, source, destination)
 	ret0, _ := ret[0].(*redis.StringCmd)
 	return ret0
 }
 
 // RPopLPush indicates an expected call of RPopLPush.
-func (mr *MockRedisClientMockRecorder) RPopLPush(source, destination interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) RPopLPush(ctx, source, destination any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RPopLPush", reflect.TypeOf((*MockRedisClient)(nil).RPopLPush), source, destination)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RPopLPush", reflect.TypeOf((*MockRedisClient)(nil).RPopLPush), ctx, source, destination)
 }
 
 // RPush mocks base method.
-func (m *MockRedisClient) RPush(key string, values ...interface{}) *redis.IntCmd {
+func (m *MockRedisClient) RPush(ctx context.Context, key string, values ...any) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{key}
+	varargs := []any{ctx, key}
 	for _, a := range values {
 		varargs = append(varargs, a)
 	}
@@ -376,16 +382,16 @@ func (m *MockRedisClient) RPush(key string, values ...interface{}) *redis.IntCmd
 }
 
 // RPush indicates an expected call of RPush.
-func (mr *MockRedisClientMockRecorder) RPush(key interface{}, values ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) RPush(ctx, key any, values ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{key}, values...)
+	varargs := append([]any{ctx, key}, values...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "RPush", reflect.TypeOf((*MockRedisClient)(nil).RPush), varargs...)
 }
 
 // SAdd mocks base method.
-func (m *MockRedisClient) SAdd(key string, members ...interface{}) *redis.IntCmd {
+func (m *MockRedisClient) SAdd(ctx context.Context, key string, members ...any) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{key}
+	varargs := []any{ctx, key}
 	for _, a := range members {
 		varargs = append(varargs, a)
 	}
@@ -395,72 +401,72 @@ func (m *MockRedisClient) SAdd(key string, members ...interface{}) *redis.IntCmd
 }
 
 // SAdd indicates an expected call of SAdd.
-func (mr *MockRedisClientMockRecorder) SAdd(key interface{}, members ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) SAdd(ctx, key any, members ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{key}, members...)
+	varargs := append([]any{ctx, key}, members...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SAdd", reflect.TypeOf((*MockRedisClient)(nil).SAdd), varargs...)
 }
 
 // SCard mocks base method.
-func (m *MockRedisClient) SCard(key string) *redis.IntCmd {
+func (m *MockRedisClient) SCard(ctx context.Context, key string) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SCard", key)
+	ret := m.ctrl.Call(m, "SCard", ctx, key)
 	ret0, _ := ret[0].(*redis.IntCmd)
 	return ret0
 }
 
 // SCard indicates an expected call of SCard.
-func (mr *MockRedisClientMockRecorder) SCard(key interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) SCard(ctx, key any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SCard", reflect.TypeOf((*MockRedisClient)(nil).SCard), key)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SCard", reflect.TypeOf((*MockRedisClient)(nil).SCard), ctx, key)
 }
 
 // SIsMember mocks base method.
-func (m *MockRedisClient) SIsMember(key string, member interface{}) *redis.BoolCmd {
+func (m *MockRedisClient) SIsMember(ctx context.Context, key string, member any) *redis.BoolCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SIsMember", key, member)
+	ret := m.ctrl.Call(m, "SIsMember", ctx, key, member)
 	ret0, _ := ret[0].(*redis.BoolCmd)
 	return ret0
 }
 
 // SIsMember indicates an expected call of SIsMember.
-func (mr *MockRedisClientMockRecorder) SIsMember(key, member interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) SIsMember(ctx, key, member any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SIsMember", reflect.TypeOf((*MockRedisClient)(nil).SIsMember), key, member)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SIsMember", reflect.TypeOf((*MockRedisClient)(nil).SIsMember), ctx, key, member)
 }
 
 // SMembers mocks base method.
-func (m *MockRedisClient) SMembers(key string) *redis.StringSliceCmd {
+func (m *MockRedisClient) SMembers(ctx context.Context, key string) *redis.StringSliceCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SMembers", key)
+	ret := m.ctrl.Call(m, "SMembers", ctx, key)
 	ret0, _ := ret[0].(*redis.StringSliceCmd)
 	return ret0
 }
 
 // SMembers indicates an expected call of SMembers.
-func (mr *MockRedisClientMockRecorder) SMembers(key interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) SMembers(ctx, key any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SMembers", reflect.TypeOf((*MockRedisClient)(nil).SMembers), key)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SMembers", reflect.TypeOf((*MockRedisClient)(nil).SMembers), ctx, key)
 }
 
 // SPopN mocks base method.
-func (m *MockRedisClient) SPopN(key string, count int64) *redis.StringSliceCmd {
+func (m *MockRedisClient) SPopN(ctx context.Context, key string, count int64) *redis.StringSliceCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SPopN", key, count)
+	ret := m.ctrl.Call(m, "SPopN", ctx, key, count)
 	ret0, _ := ret[0].(*redis.StringSliceCmd)
 	return ret0
 }
 
 // SPopN indicates an expected call of SPopN.
-func (mr *MockRedisClientMockRecorder) SPopN(key, count interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) SPopN(ctx, key, count any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SPopN", reflect.TypeOf((*MockRedisClient)(nil).SPopN), key, count)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SPopN", reflect.TypeOf((*MockRedisClient)(nil).SPopN), ctx, key, count)
 }
 
 // SRem mocks base method.
-func (m *MockRedisClient) SRem(key string, members ...interface{}) *redis.IntCmd {
+func (m *MockRedisClient) SRem(ctx context.Context, key string, members ...any) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{key}
+	varargs := []any{ctx, key}
 	for _, a := range members {
 		varargs = append(varargs, a)
 	}
@@ -470,16 +476,16 @@ func (m *MockRedisClient) SRem(key string, members ...interface{}) *redis.IntCmd
 }
 
 // SRem indicates an expected call of SRem.
-func (mr *MockRedisClientMockRecorder) SRem(key interface{}, members ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) SRem(ctx, key any, members ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{key}, members...)
+	varargs := append([]any{ctx, key}, members...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SRem", reflect.TypeOf((*MockRedisClient)(nil).SRem), varargs...)
 }
 
 // ScriptExists mocks base method.
-func (m *MockRedisClient) ScriptExists(scripts ...string) *redis.BoolSliceCmd {
+func (m *MockRedisClient) ScriptExists(ctx context.Context, scripts ...string) *redis.BoolSliceCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{}
+	varargs := []any{ctx}
 	for _, a := range scripts {
 		varargs = append(varargs, a)
 	}
@@ -489,65 +495,66 @@ func (m *MockRedisClient) ScriptExists(scripts ...string) *redis.BoolSliceCmd {
 }
 
 // ScriptExists indicates an expected call of ScriptExists.
-func (mr *MockRedisClientMockRecorder) ScriptExists(scripts ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ScriptExists(ctx any, scripts ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ScriptExists", reflect.TypeOf((*MockRedisClient)(nil).ScriptExists), scripts...)
+	varargs := append([]any{ctx}, scripts...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ScriptExists", reflect.TypeOf((*MockRedisClient)(nil).ScriptExists), varargs...)
 }
 
 // ScriptLoad mocks base method.
-func (m *MockRedisClient) ScriptLoad(script string) *redis.StringCmd {
+func (m *MockRedisClient) ScriptLoad(ctx context.Context, script string) *redis.StringCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ScriptLoad", script)
+	ret := m.ctrl.Call(m, "ScriptLoad", ctx, script)
 	ret0, _ := ret[0].(*redis.StringCmd)
 	return ret0
 }
 
 // ScriptLoad indicates an expected call of ScriptLoad.
-func (mr *MockRedisClientMockRecorder) ScriptLoad(script interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ScriptLoad(ctx, script any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ScriptLoad", reflect.TypeOf((*MockRedisClient)(nil).ScriptLoad), script)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ScriptLoad", reflect.TypeOf((*MockRedisClient)(nil).ScriptLoad), ctx, script)
 }
 
 // Set mocks base method.
-func (m *MockRedisClient) Set(key string, value interface{}, expiration time.Duration) *redis.StatusCmd {
+func (m *MockRedisClient) Set(ctx context.Context, key string, value any, expiration time.Duration) *redis.StatusCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Set", key, value, expiration)
+	ret := m.ctrl.Call(m, "Set", ctx, key, value, expiration)
 	ret0, _ := ret[0].(*redis.StatusCmd)
 	return ret0
 }
 
 // Set indicates an expected call of Set.
-func (mr *MockRedisClientMockRecorder) Set(key, value, expiration interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) Set(ctx, key, value, expiration any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Set", reflect.TypeOf((*MockRedisClient)(nil).Set), key, value, expiration)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Set", reflect.TypeOf((*MockRedisClient)(nil).Set), ctx, key, value, expiration)
 }
 
 // SetNX mocks base method.
-func (m *MockRedisClient) SetNX(key string, value interface{}, expiration time.Duration) *redis.BoolCmd {
+func (m *MockRedisClient) SetNX(ctx context.Context, key string, value any, expiration time.Duration) *redis.BoolCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SetNX", key, value, expiration)
+	ret := m.ctrl.Call(m, "SetNX", ctx, key, value, expiration)
 	ret0, _ := ret[0].(*redis.BoolCmd)
 	return ret0
 }
 
 // SetNX indicates an expected call of SetNX.
-func (mr *MockRedisClientMockRecorder) SetNX(key, value, expiration interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) SetNX(ctx, key, value, expiration any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetNX", reflect.TypeOf((*MockRedisClient)(nil).SetNX), key, value, expiration)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SetNX", reflect.TypeOf((*MockRedisClient)(nil).SetNX), ctx, key, value, expiration)
 }
 
 // TTL mocks base method.
-func (m *MockRedisClient) TTL(key string) *redis.DurationCmd {
+func (m *MockRedisClient) TTL(ctx context.Context, key string) *redis.DurationCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "TTL", key)
+	ret := m.ctrl.Call(m, "TTL", ctx, key)
 	ret0, _ := ret[0].(*redis.DurationCmd)
 	return ret0
 }
 
 // TTL indicates an expected call of TTL.
-func (mr *MockRedisClientMockRecorder) TTL(key interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) TTL(ctx, key any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "TTL", reflect.TypeOf((*MockRedisClient)(nil).TTL), key)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "TTL", reflect.TypeOf((*MockRedisClient)(nil).TTL), ctx, key)
 }
 
 // TxPipeline mocks base method.
@@ -564,24 +571,10 @@ func (mr *MockRedisClientMockRecorder) TxPipeline() *gomock.Call {
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "TxPipeline", reflect.TypeOf((*MockRedisClient)(nil).TxPipeline))
 }
 
-// WithContext mocks base method.
-func (m *MockRedisClient) WithContext(arg0 context.Context) *redis.Client {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "WithContext", arg0)
-	ret0, _ := ret[0].(*redis.Client)
-	return ret0
-}
-
-// WithContext indicates an expected call of WithContext.
-func (mr *MockRedisClientMockRecorder) WithContext(arg0 interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "WithContext", reflect.TypeOf((*MockRedisClient)(nil).WithContext), arg0)
-}
-
 // ZAdd mocks base method.
-func (m *MockRedisClient) ZAdd(key string, members ...redis.Z) *redis.IntCmd {
+func (m *MockRedisClient) ZAdd(ctx context.Context, key string, members ...redis.Z) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{key}
+	varargs := []any{ctx, key}
 	for _, a := range members {
 		varargs = append(varargs, a)
 	}
@@ -591,86 +584,86 @@ func (m *MockRedisClient) ZAdd(key string, members ...redis.Z) *redis.IntCmd {
 }
 
 // ZAdd indicates an expected call of ZAdd.
-func (mr *MockRedisClientMockRecorder) ZAdd(key interface{}, members ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZAdd(ctx, key any, members ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{key}, members...)
+	varargs := append([]any{ctx, key}, members...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZAdd", reflect.TypeOf((*MockRedisClient)(nil).ZAdd), varargs...)
 }
 
 // ZCard mocks base method.
-func (m *MockRedisClient) ZCard(key string) *redis.IntCmd {
+func (m *MockRedisClient) ZCard(ctx context.Context, key string) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZCard", key)
+	ret := m.ctrl.Call(m, "ZCard", ctx, key)
 	ret0, _ := ret[0].(*redis.IntCmd)
 	return ret0
 }
 
 // ZCard indicates an expected call of ZCard.
-func (mr *MockRedisClientMockRecorder) ZCard(key interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZCard(ctx, key any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZCard", reflect.TypeOf((*MockRedisClient)(nil).ZCard), key)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZCard", reflect.TypeOf((*MockRedisClient)(nil).ZCard), ctx, key)
 }
 
 // ZRangeByScore mocks base method.
-func (m *MockRedisClient) ZRangeByScore(key string, opt redis.ZRangeBy) *redis.StringSliceCmd {
+func (m *MockRedisClient) ZRangeByScore(ctx context.Context, key string, opt *redis.ZRangeBy) *redis.StringSliceCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZRangeByScore", key, opt)
+	ret := m.ctrl.Call(m, "ZRangeByScore", ctx, key, opt)
 	ret0, _ := ret[0].(*redis.StringSliceCmd)
 	return ret0
 }
 
 // ZRangeByScore indicates an expected call of ZRangeByScore.
-func (mr *MockRedisClientMockRecorder) ZRangeByScore(key, opt interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZRangeByScore(ctx, key, opt any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRangeByScore", reflect.TypeOf((*MockRedisClient)(nil).ZRangeByScore), key, opt)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRangeByScore", reflect.TypeOf((*MockRedisClient)(nil).ZRangeByScore), ctx, key, opt)
 }
 
 // ZRangeByScoreWithScores mocks base method.
-func (m *MockRedisClient) ZRangeByScoreWithScores(key string, opt redis.ZRangeBy) *redis.ZSliceCmd {
+func (m *MockRedisClient) ZRangeByScoreWithScores(ctx context.Context, key string, opt *redis.ZRangeBy) *redis.ZSliceCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZRangeByScoreWithScores", key, opt)
+	ret := m.ctrl.Call(m, "ZRangeByScoreWithScores", ctx, key, opt)
 	ret0, _ := ret[0].(*redis.ZSliceCmd)
 	return ret0
 }
 
 // ZRangeByScoreWithScores indicates an expected call of ZRangeByScoreWithScores.
-func (mr *MockRedisClientMockRecorder) ZRangeByScoreWithScores(key, opt interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZRangeByScoreWithScores(ctx, key, opt any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRangeByScoreWithScores", reflect.TypeOf((*MockRedisClient)(nil).ZRangeByScoreWithScores), key, opt)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRangeByScoreWithScores", reflect.TypeOf((*MockRedisClient)(nil).ZRangeByScoreWithScores), ctx, key, opt)
 }
 
 // ZRangeWithScores mocks base method.
-func (m *MockRedisClient) ZRangeWithScores(key string, start, stop int64) *redis.ZSliceCmd {
+func (m *MockRedisClient) ZRangeWithScores(ctx context.Context, key string, start, stop int64) *redis.ZSliceCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZRangeWithScores", key, start, stop)
+	ret := m.ctrl.Call(m, "ZRangeWithScores", ctx, key, start, stop)
 	ret0, _ := ret[0].(*redis.ZSliceCmd)
 	return ret0
 }
 
 // ZRangeWithScores indicates an expected call of ZRangeWithScores.
-func (mr *MockRedisClientMockRecorder) ZRangeWithScores(key, start, stop interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZRangeWithScores(ctx, key, start, stop any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRangeWithScores", reflect.TypeOf((*MockRedisClient)(nil).ZRangeWithScores), key, start, stop)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRangeWithScores", reflect.TypeOf((*MockRedisClient)(nil).ZRangeWithScores), ctx, key, start, stop)
 }
 
 // ZRank mocks base method.
-func (m *MockRedisClient) ZRank(key, member string) *redis.IntCmd {
+func (m *MockRedisClient) ZRank(ctx context.Context, key, member string) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZRank", key, member)
+	ret := m.ctrl.Call(m, "ZRank", ctx, key, member)
 	ret0, _ := ret[0].(*redis.IntCmd)
 	return ret0
 }
 
 // ZRank indicates an expected call of ZRank.
-func (mr *MockRedisClientMockRecorder) ZRank(key, member interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZRank(ctx, key, member any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRank", reflect.TypeOf((*MockRedisClient)(nil).ZRank), key, member)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRank", reflect.TypeOf((*MockRedisClient)(nil).ZRank), ctx, key, member)
 }
 
 // ZRem mocks base method.
-func (m *MockRedisClient) ZRem(key string, members ...interface{}) *redis.IntCmd {
+func (m *MockRedisClient) ZRem(ctx context.Context, key string, members ...any) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	varargs := []interface{}{key}
+	varargs := []any{ctx, key}
 	for _, a := range members {
 		varargs = append(varargs, a)
 	}
@@ -680,78 +673,78 @@ func (m *MockRedisClient) ZRem(key string, members ...interface{}) *redis.IntCmd
 }
 
 // ZRem indicates an expected call of ZRem.
-func (mr *MockRedisClientMockRecorder) ZRem(key interface{}, members ...interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZRem(ctx, key any, members ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	varargs := append([]interface{}{key}, members...)
+	varargs := append([]any{ctx, key}, members...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRem", reflect.TypeOf((*MockRedisClient)(nil).ZRem), varargs...)
 }
 
 // ZRevRangeByScore mocks base method.
-func (m *MockRedisClient) ZRevRangeByScore(key string, opt redis.ZRangeBy) *redis.StringSliceCmd {
+func (m *MockRedisClient) ZRevRangeByScore(ctx context.Context, key string, opt *redis.ZRangeBy) *redis.StringSliceCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZRevRangeByScore", key, opt)
+	ret := m.ctrl.Call(m, "ZRevRangeByScore", ctx, key, opt)
 	ret0, _ := ret[0].(*redis.StringSliceCmd)
 	return ret0
 }
 
 // ZRevRangeByScore indicates an expected call of ZRevRangeByScore.
-func (mr *MockRedisClientMockRecorder) ZRevRangeByScore(key, opt interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZRevRangeByScore(ctx, key, opt any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRevRangeByScore", reflect.TypeOf((*MockRedisClient)(nil).ZRevRangeByScore), key, opt)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRevRangeByScore", reflect.TypeOf((*MockRedisClient)(nil).ZRevRangeByScore), ctx, key, opt)
 }
 
 // ZRevRangeByScoreWithScores mocks base method.
-func (m *MockRedisClient) ZRevRangeByScoreWithScores(key string, opt redis.ZRangeBy) *redis.ZSliceCmd {
+func (m *MockRedisClient) ZRevRangeByScoreWithScores(ctx context.Context, key string, opt *redis.ZRangeBy) *redis.ZSliceCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZRevRangeByScoreWithScores", key, opt)
+	ret := m.ctrl.Call(m, "ZRevRangeByScoreWithScores", ctx, key, opt)
 	ret0, _ := ret[0].(*redis.ZSliceCmd)
 	return ret0
 }
 
 // ZRevRangeByScoreWithScores indicates an expected call of ZRevRangeByScoreWithScores.
-func (mr *MockRedisClientMockRecorder) ZRevRangeByScoreWithScores(key, opt interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZRevRangeByScoreWithScores(ctx, key, opt any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRevRangeByScoreWithScores", reflect.TypeOf((*MockRedisClient)(nil).ZRevRangeByScoreWithScores), key, opt)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRevRangeByScoreWithScores", reflect.TypeOf((*MockRedisClient)(nil).ZRevRangeByScoreWithScores), ctx, key, opt)
 }
 
 // ZRevRangeWithScores mocks base method.
-func (m *MockRedisClient) ZRevRangeWithScores(key string, start, stop int64) *redis.ZSliceCmd {
+func (m *MockRedisClient) ZRevRangeWithScores(ctx context.Context, key string, start, stop int64) *redis.ZSliceCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZRevRangeWithScores", key, start, stop)
+	ret := m.ctrl.Call(m, "ZRevRangeWithScores", ctx, key, start, stop)
 	ret0, _ := ret[0].(*redis.ZSliceCmd)
 	return ret0
 }
 
 // ZRevRangeWithScores indicates an expected call of ZRevRangeWithScores.
-func (mr *MockRedisClientMockRecorder) ZRevRangeWithScores(key, start, stop interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZRevRangeWithScores(ctx, key, start, stop any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRevRangeWithScores", reflect.TypeOf((*MockRedisClient)(nil).ZRevRangeWithScores), key, start, stop)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRevRangeWithScores", reflect.TypeOf((*MockRedisClient)(nil).ZRevRangeWithScores), ctx, key, start, stop)
 }
 
 // ZRevRank mocks base method.
-func (m *MockRedisClient) ZRevRank(key, member string) *redis.IntCmd {
+func (m *MockRedisClient) ZRevRank(ctx context.Context, key, member string) *redis.IntCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZRevRank", key, member)
+	ret := m.ctrl.Call(m, "ZRevRank", ctx, key, member)
 	ret0, _ := ret[0].(*redis.IntCmd)
 	return ret0
 }
 
 // ZRevRank indicates an expected call of ZRevRank.
-func (mr *MockRedisClientMockRecorder) ZRevRank(key, member interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZRevRank(ctx, key, member any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRevRank", reflect.TypeOf((*MockRedisClient)(nil).ZRevRank), key, member)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZRevRank", reflect.TypeOf((*MockRedisClient)(nil).ZRevRank), ctx, key, member)
 }
 
 // ZScore mocks base method.
-func (m *MockRedisClient) ZScore(key, member string) *redis.FloatCmd {
+func (m *MockRedisClient) ZScore(ctx context.Context, key, member string) *redis.FloatCmd {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "ZScore", key, member)
+	ret := m.ctrl.Call(m, "ZScore", ctx, key, member)
 	ret0, _ := ret[0].(*redis.FloatCmd)
 	return ret0
 }
 
 // ZScore indicates an expected call of ZScore.
-func (mr *MockRedisClientMockRecorder) ZScore(key, member interface{}) *gomock.Call {
+func (mr *MockRedisClientMockRecorder) ZScore(ctx, key, member any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZScore", reflect.TypeOf((*MockRedisClient)(nil).ZScore), key, member)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ZScore", reflect.TypeOf((*MockRedisClient)(nil).ZScore), ctx, key, member)
 }
