@@ -4,7 +4,7 @@
 package mocks
 
 import (
-	gomock "github.com/golang/mock/gomock"
+	gomock "go.uber.org/mock/gomock"
 	time "time"
 )
 
