@@ -8,8 +8,8 @@
 package dogstatsd_test
 
 import (
-	"github.com/golang/mock/gomock"
-	. "github.com/onsi/ginkgo"
+	"go.uber.org/mock/gomock"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 
 	"testing"
