@@ -27,7 +27,7 @@ import (
 	"time"
 
 	"github.com/confluentinc/confluent-kafka-go/kafka"
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/sirupsen/logrus"
 	"github.com/sirupsen/logrus/hooks/test"
@@ -375,7 +375,7 @@ var _ = Describe("Kafka Extension", func() {
 					&kafka.Message{
 						TopicPartition: kafka.TopicPartition{
 							Topic:     &client.Topics[0],
-							Partition: kafka.PartitionAny,
+							Partition: int32(kafka.PartitionAny),
 						},
 						Value: []byte("Hello Go!")},
 					nil,
