@@ -1,6 +1,6 @@
 package pg
 
-import "github.com/go-pg/pg/orm"
+import "github.com/go-pg/pg/v10/orm"
 
 // TestResult struct
 type TestResult struct {
