@@ -23,7 +23,7 @@
 package regex_test
 
 import (
-	. "github.com/onsi/ginkgo"
+	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	"github.com/topfreegames/extensions/v9/regex"
 )
