@@ -1,93 +1,95 @@
 // Code generated by MockGen. DO NOT EDIT.
-// Source: interfaces/s3.go
+// Source: s3/interfaces/s3.go
 
-// Package mock_interfaces is a generated GoMock package.
-package mock_interfaces
+// Package mocks is a generated GoMock package.
+package mocks
 
 import (
-	s3 "github.com/aws/aws-sdk-go/service/s3"
-	gomock "github.com/golang/mock/gomock"
+	context "context"
 	http "net/http"
 	reflect "reflect"
+
+	s3 "github.com/aws/aws-sdk-go-v2/service/s3"
+	gomock "go.uber.org/mock/gomock"
 )
 
-// MockS3 is a mock of S3 interface
+// MockS3 is a mock of S3 interface.
 type MockS3 struct {
 	ctrl     *gomock.Controller
 	recorder *MockS3MockRecorder
 }
 
-// MockS3MockRecorder is the mock recorder for MockS3
+// MockS3MockRecorder is the mock recorder for MockS3.
 type MockS3MockRecorder struct {
 	mock *MockS3
 }
 
-// NewMockS3 creates a new mock instance
+// NewMockS3 creates a new mock instance.
 func NewMockS3(ctrl *gomock.Controller) *MockS3 {
 	mock := &MockS3{ctrl: ctrl}
 	mock.recorder = &MockS3MockRecorder{mock}
 	return mock
 }
 
-// EXPECT returns an object that allows the caller to indicate expected use
+// EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockS3) EXPECT() *MockS3MockRecorder {
 	return m.recorder
 }
 
-// DeleteObject mocks base method
-func (m *MockS3) DeleteObject(key string) error {
+// DeleteObject mocks base method.
+func (m *MockS3) DeleteObject(ctx context.Context, key string) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "DeleteObject", key)
+	ret := m.ctrl.Call(m, "DeleteObject", ctx, key)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
-// DeleteObject indicates an expected call of DeleteObject
-func (mr *MockS3MockRecorder) DeleteObject(key interface{}) *gomock.Call {
+// DeleteObject indicates an expected call of DeleteObject.
+func (mr *MockS3MockRecorder) DeleteObject(ctx, key interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteObject", reflect.TypeOf((*MockS3)(nil).DeleteObject), key)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteObject", reflect.TypeOf((*MockS3)(nil).DeleteObject), ctx, key)
 }
 
-// PutObjectRequest mocks base method
-func (m *MockS3) PutObjectRequest(key, acl string) (string, http.Header, error) {
+// PutObject mocks base method.
+func (m *MockS3) PutObject(ctx context.Context, key string, body *[]byte) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "PutObjectRequest", key, acl)
+	ret := m.ctrl.Call(m, "PutObject", ctx, key, body)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// PutObject indicates an expected call of PutObject.
+func (mr *MockS3MockRecorder) PutObject(ctx, key, body interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PutObject", reflect.TypeOf((*MockS3)(nil).PutObject), ctx, key, body)
+}
+
+// PutObjectInput mocks base method.
+func (m *MockS3) PutObjectInput(ctx context.Context, params *s3.PutObjectInput, body *[]byte) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "PutObjectInput", ctx, params, body)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// PutObjectInput indicates an expected call of PutObjectInput.
+func (mr *MockS3MockRecorder) PutObjectInput(ctx, params, body interface{}) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PutObjectInput", reflect.TypeOf((*MockS3)(nil).PutObjectInput), ctx, params, body)
+}
+
+// PutObjectRequest mocks base method.
+func (m *MockS3) PutObjectRequest(ctx context.Context, key, acl string) (string, http.Header, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "PutObjectRequest", ctx, key, acl)
 	ret0, _ := ret[0].(string)
 	ret1, _ := ret[1].(http.Header)
 	ret2, _ := ret[2].(error)
 	return ret0, ret1, ret2
 }
 
-// PutObjectRequest indicates an expected call of PutObjectRequest
-func (mr *MockS3MockRecorder) PutObjectRequest(key, acl interface{}) *gomock.Call {
+// PutObjectRequest indicates an expected call of PutObjectRequest.
+func (mr *MockS3MockRecorder) PutObjectRequest(ctx, key, acl interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PutObjectRequest", reflect.TypeOf((*MockS3)(nil).PutObjectRequest), key, acl)
-}
-
-// PutObject mocks base method
-func (m *MockS3) PutObject(key string, body *[]byte) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "PutObject", key, body)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// PutObject indicates an expected call of PutObject
-func (mr *MockS3MockRecorder) PutObject(key, body interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PutObject", reflect.TypeOf((*MockS3)(nil).PutObject), key, body)
-}
-
-// PutObjectInput mocks base method
-func (m *MockS3) PutObjectInput(params *s3.PutObjectInput, body *[]byte) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "PutObjectInput", params, body)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// PutObjectInput indicates an expected call of PutObjectInput
-func (mr *MockS3MockRecorder) PutObjectInput(params, body interface{}) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PutObjectInput", reflect.TypeOf((*MockS3)(nil).PutObjectInput), params, body)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "PutObjectRequest", reflect.TypeOf((*MockS3)(nil).PutObjectRequest), ctx, key, acl)
 }
